workflows/flashcard-generator/requirements.txt
llama_index
pydantic
tenacity
python-dotenv